FEAT0022: Enable deletion of question to options associations through UI
https://api.rubyonrails.org/classes/ActiveRecord/NestedAttributes/ClassMethods.html

By default you will only be able to set and update attributes on the associated model. If you want to destroy the associated model through the attributes hash, you have to enable it first using the :allow_destroy option.
add the _destroy key to the attributes hash, with a value that evaluates to true, you will destroy the associated modelS